refactor: search route details
