+++ apps/web/app/components/search-bar.jsx
@@ -1,8 +1,9 @@
 import { Link } from "@remix-run/react";
 import { useTranslation } from 'react-i18next';
 import { MagnifyingGlassIcon, ArrowRightIcon } from "@radix-ui/react-icons";
+import { generateLocalizedRoute } from "../utils/generate-route";
 
-export default function SearchBar() {
+export default function SearchBar({locale}) {
     const { t } = useTranslation();
 
     return (   
@@ -16,15 +17,30 @@ export default function SearchBar() {
               type="text"
               placeholder={t("searchBarTitle")}
               className="w-full pl-10 pr-4 py-2 rounded-lg focus:outline-none"
+              id="search-input"
             />
           </div>
-          <Link to="/stations" className="bg-blue-500 hover:bg-blue-600 transition-all text-white px-6 py-2 rounded-lg flex items-center group">
+          <button
+            type="button"
+            className="bg-blue-500 hover:bg-blue-600 transition-all text-white px-6 py-2 rounded-lg flex items-center group"
+            onClick={() => {
+              const input = document.getElementById("search-input");
+              const value = input ? input.value.trim() : "";
+              let route;
+              if (value === "") {
+                route = generateLocalizedRoute(locale, "/search");
+              } else {
+                route = generateLocalizedRoute(locale, `/search?q=${encodeURIComponent(value)}`);
+              }
+              window.location.href = route;
+            }}
+          >
             <img src="/assets/equalizer.svg" alt="eq" className="mr-2"/>
             <span>{t("stations")}</span>
             <span className="w-0 overflow-hidden group-hover:w-5 transition-all duration-300 ease-in-out">
               <ArrowRightIcon className="ml-1 mt-0.5 w-4 h-4"/>
             </span>
-          </Link>
+          </button>
         </div>
       </div>
     );
